+++ app/src/main/res/layout/fragment_home_fragment.xml
@@ -8,17 +8,18 @@
     tools:context=".Home_fragment">
 
     <!-- TODO: Update blank fragment layout -->
+
     <TextView
         android:id="@+id/text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
-        android:layout_marginTop="70dp"
+        android:layout_marginTop="50dp"
+        android:gravity="center_horizontal"
         android:text="@string/weather"
-        android:textSize="40sp"
         android:textColor="@color/white"
-        android:textStyle="bold"
-        android:gravity="center_horizontal"/>
+        android:textSize="40sp"
+        android:textStyle="bold" />
 
     <EditText
         android:id="@+id/tx_mos"
@@ -93,7 +94,7 @@
 
             <TextView
                 android:id="@+id/textView4"
-                android:layout_width="0dp"
+                android:layout_width="80dp"
                 android:layout_height="wrap_content"
                 android:text="@string/humidity"
                 android:textColor="@color/purple_700"
@@ -160,7 +161,7 @@
 
             <EditText
                 android:id="@+id/textView7"
-                android:layout_width="90dp"
+                android:layout_width="100dp"
                 android:layout_height="wrap_content"
                 android:text=""
                 android:ems="10"
@@ -175,20 +176,21 @@
 
     <Button
         android:id="@+id/city"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="550dp"
-        android:text="@string/choose_city"
-        android:textSize="20sp"
-        android:textColor="@color/city"
-        android:gravity="center_horizontal"/>
-
-    <EditText
-        android:id="@+id/input"
         android:layout_width="100dp"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
-        android:layout_marginTop="600dp"
+        android:layout_marginTop="550dp"
+        android:background="@color/colorPrimary"
+        android:text="@string/choose_city"
+        android:textColor="@color/white"
+        android:textSize="20sp" />
+
+    <EditText
+        android:id="@+id/input"
+        android:layout_width="150dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginTop="660dp"
         android:text="" />
