+++ gradlew
@@ -11,20 +11,28 @@
 PRG="$0"
 # Need this for relative symlinks.
 while [ -h "$PRG" ] ; do
+    # shellcheck disable=SC2006
     ls=`ls -ld "$PRG"`
+    # shellcheck disable=SC2006
     link=`expr "$ls" : '.*-> \(.*\)$'`
     if expr "$link" : '/.*' > /dev/null; then
         PRG="$link"
     else
+        # shellcheck disable=SC2006
         PRG=`dirname "$PRG"`"/$link"
     fi
 done
 SAVED="`pwd`"
+# shellcheck disable=SC2164
+# shellcheck disable=SC2006
 cd "`dirname \"$PRG\"`/" >/dev/null
+# shellcheck disable=SC2006
 APP_HOME="`pwd -P`"
+# shellcheck disable=SC2164
 cd "$SAVED" >/dev/null
 
 APP_NAME="Gradle"
+# shellcheck disable=SC2006
 APP_BASE_NAME=`basename "$0"`
 
 # Add default JVM options here. You can also use JAVA_OPTS and GRADLE_OPTS to pass JVM options to this script.
@@ -49,6 +57,7 @@ cygwin=false
 msys=false
 darwin=false
 nonstop=false
+# shellcheck disable=SC2006
 case "`uname`" in
   CYGWIN* )
     cygwin=true
@@ -57,6 +66,7 @@ case "`uname`" in
     darwin=true
     ;;
   MINGW* )
+    # shellcheck disable=SC2034
     msys=true
     ;;
   NONSTOP* )
@@ -89,13 +99,17 @@ location of your Java installation."
 fi
 
 # Increase the maximum file descriptors if we can.
+# shellcheck disable=SC2166
 if [ "$cygwin" = "false" -a "$darwin" = "false" -a "$nonstop" = "false" ] ; then
+    # shellcheck disable=SC2006
+    # shellcheck disable=SC2039
     MAX_FD_LIMIT=`ulimit -H -n`
     if [ $? -eq 0 ] ; then
         if [ "$MAX_FD" = "maximum" -o "$MAX_FD" = "max" ] ; then
             MAX_FD="$MAX_FD_LIMIT"
         fi
-        ulimit -n $MAX_FD
+        # shellcheck disable=SC2039
+        ulimit -n "$MAX_FD"
         if [ $? -ne 0 ] ; then
             warn "Could not set maximum file descriptor limit: $MAX_FD"
         fi
@@ -111,11 +125,15 @@ fi
 
 # For Cygwin, switch paths to Windows format before running java
 if $cygwin ; then
+    # shellcheck disable=SC2006
     APP_HOME=`cygpath --path --mixed "$APP_HOME"`
+    # shellcheck disable=SC2006
     CLASSPATH=`cygpath --path --mixed "$CLASSPATH"`
+    # shellcheck disable=SC2006
     JAVACMD=`cygpath --unix "$JAVACMD"`
 
     # We build the pattern for arguments to be converted via cygpath
+    # shellcheck disable=SC2006
     ROOTDIRSRAW=`find -L / -maxdepth 1 -mindepth 1 -type d 2>/dev/null`
     SEP=""
     for dir in $ROOTDIRSRAW ; do
@@ -130,12 +148,22 @@ if $cygwin ; then
     # Now convert the arguments - kludge to limit ourselves to /bin/sh
     i=0
     for arg in "$@" ; do
+        # shellcheck disable=SC2006
+        # shellcheck disable=SC2196
         CHECK=`echo "$arg"|egrep -c "$OURCYGPATTERN" -`
+        # shellcheck disable=SC2006
+        # shellcheck disable=SC2196
         CHECK2=`echo "$arg"|egrep -c "^-"`                                 ### Determine if an option
 
-        if [ $CHECK -ne 0 ] && [ $CHECK2 -eq 0 ] ; then                    ### Added a condition
+        if [ "$CHECK" -ne 0 ] && [ "$CHECK2" -eq 0 ] ; then                    ### Added a condition
+            # shellcheck disable=SC2046
+            # shellcheck disable=SC2006
+            # shellcheck disable=SC2116
             eval `echo args$i`=`cygpath --path --ignore --mixed "$arg"`
         else
+            # shellcheck disable=SC2046
+            # shellcheck disable=SC2006
+            # shellcheck disable=SC2116
             eval `echo args$i`="\"$arg\""
         fi
         i=$((i+1))
@@ -162,10 +190,11 @@ save () {
 APP_ARGS=$(save "$@")
 
 # Collect all arguments for the java command, following the shell quoting and substitution rules
-eval set -- $DEFAULT_JVM_OPTS $JAVA_OPTS $GRADLE_OPTS "\"-Dorg.gradle.appname=$APP_BASE_NAME\"" -classpath "\"$CLASSPATH\"" org.gradle.wrapper.GradleWrapperMain "$APP_ARGS"
+eval set -- $DEFAULT_JVM_OPTS "$JAVA_OPTS" "$GRADLE_OPTS" "\"-Dorg.gradle.appname=$APP_BASE_NAME\"" -classpath "\"$CLASSPATH\"" org.gradle.wrapper.GradleWrapperMain "$APP_ARGS"
 
 # by default we should be in the correct project dir, but when run from Finder on Mac, the cwd is wrong
 if [ "$(uname)" = "Darwin" ] && [ "$HOME" = "$PWD" ]; then
+  # shellcheck disable=SC2164
   cd "$(dirname "$0")"
 fi
